pyproject.toml
[tool.poetry]
name = "conversational_bot_api"
version = "0.1.0"
description = ""
authors = ["mbunchsy <miguelramosjulian@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.7"
djangorestframework = "^3.15.2"
dependency-injector = "^4.46.0"
openai = "^1.68.2"
python-dotenv = "^1.0.1"
elevenlabs = "^1.54.0"
psycopg2-binary = "^2.9.9"
tenacity = "^9.0.0"
tiktoken = "^0.9.0"
pgvector = "^0.4.0"
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
db-init = "alembic:main init"
db-migrate = "alembic:main revision --autogenerate"
db-upgrade = "alembic:main upgrade head"